#8-1 민우가 알고리즘 완성

+++ final-pjt-back/movies/views.py
@@ -42,14 +42,17 @@ def movie_detail(request, movie_pk):
 def comment_detail(request, comment_pk):
     # comment = Comment.objects.get(pk=comment_pk)
     comment = get_object_or_404(Comment, pk=comment_pk)
+    user=request.user
 
     if request.method == 'GET':
         serializer = CommentSerializer(comment)
         return Response(serializer.data)
-
     elif request.method == 'DELETE':
-        comment.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        # print(user.pk)
+        # print(comment.user_id)
+        if user.pk == comment.user_id:
+            comment.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
 
     elif request.method == 'PUT':
         serializer = CommentSerializer(comment, data=request.data)
